requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
requests